almost rounds

+++ databases/mb-db/src/main/resources/db/migration/V002__create_battles.sql
@@ -1,6 +1,6 @@
 create table battles
 (
     id                serial primary key,
-    winner_id         int null, foreign key (winner_id) references movies(id),
+    winner_id         bigint null, foreign key (winner_id) references movies(id),
     created_at        timestamp default current_timestamp
 );
